Fixed issue with delete image by other user, password pattern and minor bug fixes.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -100,8 +100,8 @@ public class ImageController {
         String error = "Only the owner of the image can edit the image";
         User user = (User) session.getAttribute("loggeduser");
         Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
         if(image.getUser().getId() == user.getId()) {
-            String tags = convertTagsToString(image.getTags());
             model.addAttribute("image", image);
             model.addAttribute("tags", tags);
             return "images/edit";
@@ -109,7 +109,7 @@ public class ImageController {
         else {
             model.addAttribute("editError", error);
             model.addAttribute("image", image);
-            model.addAttribute("tags", image.getTags());
+            model.addAttribute("tags", tags);
             return "images/image";
         }
 
@@ -154,9 +154,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        String error = "Only the owner of the image can edit the image";
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        if(image.getUser().getId() == user.getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }else {
+            model.addAttribute("editError", error);
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/image";
+        }
     }
 
     //This method converts the image to Base64 format
